tast arc memory perf fix overflow
Just convert everything to int64 so we don't get negative numbers on
32bit boards.

BUG=b:147774334
TEST=tast run <SOME KUKUI> arc.MemoryChromeOSPerf.noarc, memory numbers
are positive.
tast run <SOME DUAL IMAGE> arc.MemoryAndroidPerf arc.MemoryShiftingPerf.vm

+++ src/chromiumos/tast/local/bundles/cros/arc/memory_shifting_perf.go
@@ -41,7 +41,7 @@ func init() {
 }
 
 // mean returns the arithmetic mean of a slice of integers.
-func mean(values []int) float64 {
+func mean(values []int64) float64 {
 	sum := 0.0
 	for _, v := range values {
 		sum += float64(v)
